% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sum.R
\name{make_sum}
\alias{make_sum}
\title{Simple function that calculates the sum}
\usage{
make_sum(a, b)
}
\arguments{
\item{a}{summand (numeric)}

\item{b}{summand (numeric)}
}
\description{
This function takes two argument (`a` and `b`) and calculates the sum of both.
}
\examples{
\dontrun{make_sum(2, 3)}
}
